Merge_sort works

+++ user/bench.h
@@ -17,7 +17,7 @@ bench_time(void)
 {
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
-	return ((uint64_t)tv.tv_sec * 1000000 + tv.tv_usec) * 1000; // ns
+	return (tv.tv_sec * 1000000 + tv.tv_usec) * 1000; // ns
 }
 
 #endif
